pyproject.toml
[tool.poetry]
name = "bookstore-services"
version = "0.1.0"
description = "Backend services para bookStore"
authors = ["Victor Mota <victor.vmt@outlook.com>"]

[tool.poetry.dependencies]
python = "3.8.12"
Flask = "^2.0.1"
gunicorn = "^20.1.0"
flask-restx = "^0.5.1"
requests = "^2.26.0"
python-dotenv = "^0.19.0"
py-money = "^0.5.0"
Flask-SQLAlchemy = "^2.5.1"
Flask-Migrate = "^3.1.0"
psycopg2-binary = "^2.9.1"

[tool.poetry.dev-dependencies]
pytest = "^5.2"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
